test: 포인트 사용 충전 테스트 추가

+++ apps/commerce-api/src/test/java/com/loopers/interfaces/point/PointV1ApiE2ETest.java
@@ -60,7 +60,7 @@ public class PointV1ApiE2ETest {
             User user = userJpaRepository.save(
                     UserFixture.createMember()
             );
-            var request = new PointV1Request.PointChargeRequest(1000);
+            var request = new PointV1Request.PointChargeRequest(1000L);
 
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
@@ -92,7 +92,7 @@ public class PointV1ApiE2ETest {
                     UserFixture.createMember()
             );
 
-            int amount = 10000;
+            Long amount = 10000L;
             Point point = pointJpaRepository.save(
                     Point.init(user.getId())
             );
@@ -124,7 +124,7 @@ public class PointV1ApiE2ETest {
         @Test
         void throwsException_whenInvalidIdIsProvided() {
             //given
-            var request = new PointV1Request.PointChargeRequest(1000);
+            var request = new PointV1Request.PointChargeRequest(1000L);
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
             headers.set("X-USER-ID", "1");
@@ -224,7 +224,7 @@ public class PointV1ApiE2ETest {
         @Test
         void throwsException_whenInvalidIdIsProvided() {
             //given
-            var request = new PointV1Request.PointChargeRequest(1000);
+            var request = new PointV1Request.PointChargeRequest(1000L);
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
             headers.set("X-USER-ID", "1");
